requirements.txt
appdirs==1.4.3
autopep8==1.6.0
backcall==0.2.0
base58==2.0.1
beautifulsoup4==4.9.0
cairocffi==1.0.0
CairoSVG==2.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cssselect2==0.3.0
decorator==5.1.0
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2.2
django-ace==1.0.11
django-ckeditor==6.1.0
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-jquery==3.1.0
django-js-asset==1.2.2
django-modeladmin-reorder==0.3.1
django-multiselectfield==0.1.12
django-taggit==1.2.0
django-widget-tweaks==1.4.5
djangorestframework==3.12.4
ecdsa==0.16.0
filelock==3.0.12
gunicorn==20.0.0
html5lib==1.0.1
idna==2.8
importlib-metadata==1.6.0
importlib-resources==1.5.0
ipykernel==4.2.2
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.1
jupyter-client==5.2.1
jupyter-core==4.9.1
lxml==4.5.0
mutagen==1.45.1
optional-django==0.3.0
parso==0.8.2
paystack==1.5.0
paystack-python==0.1
paystackapi==2.0.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==2018.11.26
prompt-toolkit==3.0.3
psycopg2==2.8.4
pycodestyle==2.8.0
pycparser==2.20
Pygments==2.10.0
PyPDF2==1.26.0
Pyphen==0.9.5
pyqt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
PyQt5-stubs==5.15.2.0
pyqt5-tools==5.15.4.3.2
python-bitcoinlib==0.11.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2019.3
pywin32==302
pyzmq==22.3.0
qrcode==6.1
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
react==4.3.0
reportlab==3.5.34
requests==2.21.0
simplejson==3.17.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
tinycss2==1.0.2
toml==0.10.2
traitlets==4.3.3
urllib3==1.24.3
virtualenv==20.0.18
virtualenv-clone==0.5.4
wcwidth==0.2.5
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.0.1
xhtml2pdf==0.2.4
zipp==3.1.0
